image: Add a common compression type detection function.
commit 155d6a3575470c1a735b8cf368d9e987930910a8 upstream

Currently, there is no method that can detect compression types
given a file. This is very useful where a compressed kernel image
is loaded directly to the memory.

Inspect initial few bytes to figure out compression type of the
image. It will be used in booti method for now but can be reused
any other function in future as well.

Change-Id: Ibb49a23e62021db26a1b0f167075eb2bd9cd9b76
Signed-off-by: Atish Patra <atish.patra@wdc.com>
Reviewed-by: Tom Rini <trini@konsulko.com>
Reviewed-on: https://sc-debu-git.synaptics.com/gerrit/222393
Reviewed-by: Jisheng Zhang <Jisheng.Zhang@synaptics.com>
Reviewed-by: Terry Zhou <Terry.Zhou@synaptics.com>