image: Add compressed Image parsing support in booti.
commit 414c34ed555b8ce5c260cf641261ecf45beca251 upstream

Add compressed Image parsing support so that booti can parse both
flat and compressed Image to boot Linux. Currently, it is difficult
to calculate a safe address for every board where the compressed
image can be decompressed. It is also not possible to figure out the
size of the compressed file as well. Thus, user need to set two
additional environment variables kernel_comp_addr_r and filesize to
make this work.

Following compression methods are supported for now.
lzma, lzo, bzip2, gzip.

lz4 support is not added as ARM64 kernel generates a lz4 compressed
image with legacy header which U-Boot doesn't know how to parse and
decompress.

Tested on HiFive Unleashed and Qemu for RISC-V.
Tested on Qemu for ARM64.

Change-Id: I86b3843ae5e30871baa5b231a05534047058598d
Signed-off-by: Atish Patra <atish.patra@wdc.com>
Reviewed-by: Tom Rini <trini@konsulko.com>
[trini: Fix minor rST formatting problems]
Signed-off-by: Tom Rini <trini@konsulko.com>
Reviewed-on: https://sc-debu-git.synaptics.com/gerrit/222398
Reviewed-by: Jisheng Zhang <Jisheng.Zhang@synaptics.com>
Reviewed-by: Terry Zhou <Terry.Zhou@synaptics.com>